store the episode duration in seconds instead of string
The netlify-cms dependency js-yaml changed the way some strings are stored as frontmatter. In some cases strings with colons are not escaped if they do not need to be escaped. In turn, Gatsby parses the markdown and converts the unquoted string to seconds (ie: 10:20 becomes 620) while other durations surrounded by quotes are parsed as strings and not converted.

The iTunes RSS tag itunes:duration accepts various formats for the duration but recommends using seconds.

So it seems reasonable to convert duration to seconds.